Add OceanIOStep for reading/writing ocean model files
This step is appropriate for most ocean steps that are not
OceanModelSteps, since often variables should be translated to
or from the native model names to MPAS-Ocean names for use in
Polaris
