+++ polaris/ocean/model/__init__.py
@@ -1,2 +1,3 @@
+from polaris.ocean.model.ocean_io_step import OceanIOStep
 from polaris.ocean.model.ocean_model_step import OceanModelStep
 from polaris.ocean.model.time import get_time_interval_string
